Cursor provider: dedupe AsyncQueue, follow model-id normalization convention
- Extract the duplicated AsyncQueue<T> into src/server/async-queue.ts and
  import it in both codex-app-server.ts and cursor-cli.ts (was a verbatim copy).
- Add normalizeCursorModelId (mirrors normalizeClaude/CodexModelId) and use it
  in both normalizeCursorPreference functions instead of hardcoding "composer-2.5".
